CPU now passes most common tests, switched back to int for convenience
nestest.nes - 13/13 (Official OPCodes)
NEStress.nes - CPU 48/48
branch_timing - 3/3 PASSED
cpu_dummy_reads.nes - PASSED
test_cpu_exec_space_ppuio.nes - PASSED
cpu_reset - 2/2 PASSED
cpu_timing_test.nes - PASSED
dma_2007_write.nes - PASSED
instr_misc - 3/4 PASSED (no APU)
oam_read.nes - PASSED
Donkey Kong, Mario Bros and Ice Climbers 100% playable (and probably all Mapper 000 games)
Other Mappers still don't work properly (or at all)

+++ src/core/cartridge/Header.java
@@ -10,14 +10,14 @@ import java.io.IOException;
  */
 class Header {
 
-    final byte prg_rom_chunks;
-    final byte chr_rom_chunks;
-    final byte mapper1;
-    final byte mapper2;
+    final int prg_rom_chunks;
+    final int chr_rom_chunks;
+    final int mapper1;
+    final int mapper2;
     private final char[] name;
-    private final byte prg_ram_size;
-    private final byte tv_system1;
-    private final byte tv_system2;
+    private final int prg_ram_size;
+    private final int tv_system1;
+    private final int tv_system2;
 
     /**
      * Load the header from the FileReader
